backend/server/requirements.txt
mcp[cli]
uvicorn
fastapi
python-dotenv
pymongo